# docs/swagger.yaml
swagger: '2.0'
info:
  title: Living Progress - Build - Mentor Me
  description: Living Progress - Build - Mentor Me API.
  version: "1.0.0"
# the sample made up domain of the service
host: api.mentorme.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
securityDefinitions:
  basicAuth:
    type: basic
# HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
basePath: /v1
produces:
  - application/json
paths:
...
  /users/{id}:
    get:
      tags:
        - User
      summary: get User
      description: This endpoint is responsible for getting User
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the User
          schema:
            $ref: '#/definitions/User'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - User
      summary: update User
      description: This endpoint is responsible for updating User
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
        - name: user
          in: body
          description: The User to update
          schema:
            $ref: '#/definitions/User'
      responses:
        200:
          description: The response containing the User
          schema:
            $ref: '#/definitions/User'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
        - User
      summary: delete User
      description: This endpoint is responsible for deleting User
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /users:
    get:
      tags:
        - User
      summary: search for entities
      description: This endpoint is responsible for searching by attribute
      security:
        - basicAuth: []
      parameters:
        - name: name
          in: query
          description: The attribute Name
          type: string
          required: false
        - name: pageNumber
          in: query
          description: The page number
          type: number
          format: int32
          required: false
        - name: pageSize
          in: query
          description: The page size
          type: number
          format: int32
          required: false
        - name: sortColumn
          in: query
          description: The sort column
          type: string
          required: true
        - name: sortOrder
          in: query
          description: The sort order
          type: string
          required: false
      responses:
        200:
          description: The response containing the Users
          schema:
            $ref: '#/definitions/UserSearchResult'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - User
      summary: create User
      description: This endpoint is responsible for creating User
      security:
        - basicAuth: []
      parameters:
        - name: user
          in: body
          description: The User to create
          schema:
            $ref: '#/definitions/UserCreateBody'
      responses:
        200:
          description: The response containing the User
          schema:
            $ref: '#/definitions/User'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /users/forgotPassword:
    put:
      tags:
        - User
      summary: initiate forgot password
      description: This endpoint is responsible for forgot password functionality
      security:
        - basicAuth: []
      parameters:
        - name: email
          in: query
          description: The email
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /users/updatePassword:
    put:
      tags:
        - User
      summary: initiate forgot password
      description: This endpoint is responsible for forgot password functionality
      security:
        - basicAuth: []
      parameters:
        - name: newPassword
          in: body
          description: The new password.
          schema:
             $ref: '#/definitions/NewPassword'
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  UserCreateBody:
    type: object
    properties:
      username:
        type: string
        description: The username.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      password:
        type: string
        description: The password.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
  User:
    type: object
    properties:
      id:
        type: number
        format: integer
        description: The id.
      username:
        type: string
        description: The username.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      password:
        type: string
        description: The password (present only in PUT and POST requests, not in GET).
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
  UserSearchResult:
    type: object
    properties:
      total:
        type: number
        format: integer
        description: The total.
      totalPages:
        type: number
        format: integer
        description: The total pages.
      entities:
        type: array
        items:
          $ref: '#/definitions/User'
  UserRole:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
  NewPassword:
      type: object
      properties:
        token:
          type: string
          format: int32
          description: The token.
        newPassword:
          type: string
          description: The new password.
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string