# src/main/java/com/livingprogress/mentorme/entities/UserRole.java
package com.livingprogress.mentorme.entities;

import lombok.Getter;
import lombok.Setter;

import javax.persistence.Entity;

/**
 * The user role.
 */
@Getter
@Setter
@Entity
public class UserRole extends LookupEntity {
}
